+++ packages/expo-image/src/Image.tsx
@@ -17,12 +17,21 @@ export class Image extends React.PureComponent<ImageProps> {
   }
 
   /**
-   * Preloads images at the given urls that can be later used in the image view.
-   * Preloaded images are always cached on the disk, so make sure to use
-   * `disk` (default) or `memory-disk` cache policy.
+   * Preloads images at the given URLs that can be later used in the image view.
+   * Preloaded images are cached to the memory and disk by default, so make sure
+   * to use `disk` (default) or `memory-disk` [cache policy](#cachepolicy).
+   * @param urls - A URL string or an array of URLs of images to prefetch.
+   * @param cachePolicy - The cache policy for prefetched images.
+   * @return A promise resolving to `true` as soon as all images have been
+   * successfully prefetched. If an image fails to be prefetched, the promise
+   * will immediately resolve to `false` regardless of whether other images have
+   * finished prefetching.
    */
-  static prefetch(urls: string | string[]): void {
-    return ExpoImageModule.prefetch(Array.isArray(urls) ? urls : [urls]);
+  static async prefetch(
+    urls: string | string[],
+    cachePolicy: 'memory-disk' | 'memory' = 'memory-disk'
+  ): Promise<boolean> {
+    return ExpoImageModule.prefetch(Array.isArray(urls) ? urls : [urls], cachePolicy);
   }
 
   /**